+++ src/utils.ts
@@ -86,7 +86,7 @@ const flightRadarApi = new FlightRadarApi({
 
 const satelliteApi = new SatelliteServiceApi(
   new Configuration({
-    basePath: "https://satellite-api.react-flight-tracker.dev.apoorva64.com",
+    basePath: "https://satellite-api.react-flight-tracker.apoorva64.com",
   }),
 );